oox: Handle agile encryption info "reserved" field correctly
The "reserved" filed is written fter the version number major,
minor which is used to identify the encryption as agile. The
"reserved" field must always have the value 0x00000040. This
change writes the reserved filed correctly and when encryption and
when decrypting it checks the value an potentially bails out if
it desn't contain the expected value.

Change-Id: I2045dc64e0c6bbb3318384e25deef2ba8f41b94c
Reviewed-on: https://gerrit.libreoffice.org/57140
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>