related tdf#106174 writerfilter: replace broken FindParentStyleSheet
FindParentStyleSheet didn't do anything right. An empty stylename
was supposed to check currentEntry, but instead it just returned
null. The passed stylename was completely ignored if there was
a currentEntry. And on top of that, the stylename itself was
returned, not its parent. Even worse, the currentEntry properties
were ignored because in that case it DID return the parent.
Even the comments were totally bogus. Amazing...

GetPropertyFromStyleSheet() has the potential for LOTS of use
in writerfilter. I'm surprised it has hardly been used. Perhaps
the completely wrong results led people to ignore it rather than
investigate why it didn't give the expected results.

Change-Id: I89a6b02877ab7990aa4eb58e5e0b6a572f3d10e3
Reviewed-on: https://gerrit.libreoffice.org/57195
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>