tdf#118533 fix shape import: zero top margin only in first paragraph
commit 0307a62790b33ee0c02c2323a8f759e53e2035a4 fixed the
top margin import of the paragraphs of document sections, and
import of first paragraph of the text frames with beforeAutospacing,
but nullified all other paragraph top margins in frames.

Note: there is no visible margin difference in the unit test (extended
by this commit) before and after the fix, because the first paragraph
uses also afterAutospacing, resulting still 14pt paragraph space.
But the tested beforeAutospacing value of the next paragraph checks
the fix correctly.

Change-Id: I0ab3b8bbff33c5488f4b4af1ea4dabf7105103f2
Reviewed-on: https://gerrit.libreoffice.org/57231
Reviewed-by: László Németh <nemeth@numbertext.org>
Tested-by: László Németh <nemeth@numbertext.org>