tdf#63561 docx export: "clear" unused inherited tabs
If a style contains tab definitions, then the paragraph inherits
these. They are added to any tabs defined at the paragraph
level. Unwanted inherited tabs must be explicitly removed.

(TODO: LO ought to be inheriting from ALL parents, so the same
logic ought to apply to a style's parent tabs, but currently
LO does not import that way. So the proof unit test looks
different in MSO compared to LO.)

Change-Id: Ida8ed2792482655d512c753fdff8d02062d895a8
Reviewed-on: https://gerrit.libreoffice.org/57255
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>