tdf#118672 sw layout, TabOverMargin: allow using the area over the tab portion
TabOverMargin in general is about allowing the cursor to jump over a
margin if there is an explicit tab stop there.

A corner-case is what to do when there is enough content so a line break
is necessary for the characters after the tab portion. Allow using the
area up to the edge of the whole text frame (i.e. over the tab
position), this matches what Word does.

Change-Id: Ie86edf030d54fba556eee26e7ea563fb8d4fbee4
Reviewed-on: https://gerrit.libreoffice.org/57275
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins