tdf#118730 Correct ReportDesigner Object creation
This is a follow up problem to removing the old stuff
that first a SdrPage* at the SdrObjects was set *without*
the SdrObject being inserted to any SdrObjList. It works
no longer - the SdrPage which you can get now is the one
the SdrObject *is* inserted at.
Solution is to move that stuff - plus other initializations
which were done in OUnoObject::EndCreate before - to where
it belongs. This gets rid of OUnoObject::EndCreate
completely.
It makes OUnoObject::impl_setReportComponent_nothrow no
longer needed due to being used only in one place. All
initializations move to OUnoObject::CreateMediator which
anyways needs to be done when a OUnoObject got created.

Change-Id: I86f968dc6e867c5752d3c8cee1b3b2af57e467c8
Reviewed-on: https://gerrit.libreoffice.org/57976
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@cib.de>