vcl statusbar: store calculated layout in StatusBar::SetItemText()
Otherwise we calculate the layout of the text multiple times, which is a
waste of resources. This effectively reverts commit
a4ed3d9a1ffa1b51ba4352a955c950235f099fdc (tdf#115353 Status bar: no
cache in settext, 2018-03-01).

That seems to be safe, given that I can't reproduce tdf#115353 after
this change, and reading through the comments the bug was reported fixed
after both the mentioned commit and
commit 8d42909cd326cb26d8ba1fb383b5578f820c72ed (tdf#115353 layout fix,
2018-02-23). So either the first one was not necessary, or it just
worked around an underlying problem that has been fixed in the meantime.

Change-Id: I79c014e9ce41be0c06dae65b5195a23d0d88c2aa
Reviewed-on: https://gerrit.libreoffice.org/58830
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>