tdf#118058 sw, sections in tables: no split for multiple columns
The original use-case was "group a few paragraphs together" for the
split sections inside tables, i.e. it's safe to not split when the
section has multiple columns.

And the multiple columns case would mean that we don't find where to put
the follow section inside a table, resulting in a layout loop.

Change-Id: Ifab220e582439d2e757b5645f3167b55a051a379
Reviewed-on: https://gerrit.libreoffice.org/58951
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>