sw_redlinehide_2: SwUndoDelete
This is problematic because of the calls to SplitNode.
Ideally we'd want the SplitNode to create merged frames already, but
that doesn't seem to be easy to achieve; several problems with this are:

1. the redlines are only restored at the end of UndoImpl
2. even if we store another set of SwRedlineSaveDatas right before the
   Join (while preventing the first SwRedlineSaveDatas from deleting
   them), and restore them by passing a closure to SplitNode, there
   are complaints about empty redlines, and also this case isn't
   handled properly:

     f<delete start>o<redline start>o
     b<redline end>a<redline start>r
     b<redline end>a<delete end>z

So instead, let SplitNode create whatever frames it does, and fix it up
at the end manually on the start node's frames.

This necessitates delaying the creation of the frames on the moved nodes
until the end too.

Change-Id: I8ba2967659cc2ddbe6f7c40e0447d79601498ed6