properly reset nSeenInIteration when iterating cell cycles
ooo#956-3 has a somewhat complex cell cycle structure. Calc detects
cycle G40, H40, H47, H37, D40 -> G40, marks these cells as bIsIterCell,
adds them to the list of cells in RecursionHelper and starts iterating.
However, there are more cells involved, e.g. there's another cycle
D41, G41, H41, H47, H37 -> D41, which Calc doesn't detect (probably
because it partially overlaps with the detected cycle).

This means that InterpretTail() was setting nSeenInIteration for every
involved cell, but it wasn't unset, because some of the cells weren't
know to be part of the iteration. And so on the next recalc, these
cells weren't interpreted, because Interpret() aborted early because
of the stale nSeenInIteration value. And in threaded calculation, this
eventually led to hitting an assert in MaybeInterpret().

And obvious fix for this is to set nSeenInIteration only for cells
that Calc treats as being part of the iteration. Which is what this
commit does.

That's however not a complete fix. Doing a recalc with ooo#956-3 now
at least gives somewhat sensible values, but it needs repeated hard
recalcs to actually reach the correct values that converge, or the delta
change in the iteration settings need to be (needlessly) small, 1E-6,
while Excel manages with just 1E-2. So what also should be done is probably
detecting all cells involved in the cycle and treating them as such.

Change-Id: I05fd149b453363c335f1f5d5d3748354ae624613
Reviewed-on: https://gerrit.libreoffice.org/59495
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>