tdf#54819 change tracking: keep paragraph style after full deletion
of the previous paragraph(s), as in editing without change tracking.

This is a workaround for the regression introduced by #i100466#.

NOTE: We change the style of the first removed paragraph to restore
normal editing function during change tracking, too. Manual fix
of the spoiled paragraph does the same style changes in the
previously deleted paragraphs, so this commit doesn't introduce
any new oddity.

Change-Id: I9565e600ec86adc95246ff899d58de79f3c3a996
Reviewed-on: https://gerrit.libreoffice.org/59820
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>