-Werror=redundant-move (GCC 9), take two
...after 5b62a43349da6fda13fb33e0f1ec477c21daec8f "Revert
'-Werror=redundant-move'" to fix the build for GCC 8.1 again.  Turns out the
std::move can only be dropped if the compiler has a fix for CWG1579.  For GCC
that's the case starting with GCC 5.1, so the !HAVE_CXX_GWG1579_FIX case can
hopefully be removed again soon, see the mail thread starting at
<https://lists.freedesktop.org/archives/libreoffice/2018-July/080588.html>
"Compiler baselines (was: [Libreoffice-qa] minutes of ESC call ...)").

Change-Id: I3592cad7fb503db921c37e92831a34785a1054a1
Reviewed-on: https://gerrit.libreoffice.org/59741
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>