sw_redlinehide_2: call CompressRedlines later in DeleteRange*
DocumentContentOperationsManager::DeleteRange* functions should call
CompressRedlines after sw_JoinText, because otherwise if there's a join,
the FillSaveData in SwUndoDelete ctor will split redlines that contain
the selection into 2, and they won't be recombined although CanCombine
is true for them.

If you do 2 operations, then on Undo of the second, SetSaveData
will restore the previous redlines and join them into 1, but then it will
assert because it expects 2 redlines, pointlessly split.

Change-Id: I1df3f2205b4f16904f66b5af1f3b9f0ccbaf24a0