Re-introduce code to use the ICU data file in the iOS app bundle
We used to have this code snippet in the TiledLibreOffice app back in
the days. Then in LibreOfficeLight it was lost and forgotten (?).

Clearly we need to tell ICU abouyt the data that we include as a
separate file in iOS apps, and presumably any iOS app will be a
LibreOffficeKit-based one, so let's do the initialistion here.

Change-Id: Ib08dc9d7386789d10e8c53114e79d0b5beab7232