tdf#118705 Fix Column and Bar Chart data label placement
In the OOXML standard the supported data label positions in case of
Column and Bar Chart are Center (ctr), Inside End (inEnd),
Inside Base (inBase) and Outside End (outEnd). So even if the default
value is Above of the date labels placement in LibreOffice,
after the export to OOXML the "DLblPos" will be changed to outEnd.
It is not a problem, but the 0 datapoints value will be appear on the
wrong side of the axis and making itself and the the axis text both
undreadable. This patch will fix it in case of OOXML and ODF format too.

Change-Id: I04d68ee79b560cf144e6290271f2341355a5b089
Reviewed-on: https://gerrit.libreoffice.org/60477
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>
Tested-by: László Németh <nemeth@numbertext.org>