tdf#41542 sw cleanup: m_bBorderDist was always true
This was forced to True in LO 5.4, leaving the variable intact
in case any problems arose that would make it easier to revert.
But now it is time to clean it up and remove the unnecessary
logic.

I didn't change the indenting because I don't like obscuring
the significant changes. I can do that in a followup commit.

Change-Id: Iab04a6a8f4ba8ff7151cf7941f655173136258eb
Reviewed-on: https://gerrit.libreoffice.org/61270
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>