tdf#117721 draw ui: modify Optimize/Distribute RowHeight
Optimize row height: Adjust the height of the selected rows
to match the height of the tallest row in the selection
(fit to content), without shrinking the table.

(Optimize rows is a new feature for Draw. My initial
implementation was replaced by setMinimalRowHeight.
This option is now the same as minimizing row height and then
distributing rows evenly except that it adds the benefit of
preventing the table from shrinking.)

Distribute rows evenly: Adjusts the height of the selected rows
to match the height of the tallest row in the selection
(regardless of the content), causing the table to grow.

(Previously, Distribute rows worked like Optimize now does,
but the documentation indicates that it should work this way.)

Change-Id: I49b9f9b5d1f9d3e8d2267ba0d49a9901585b4095
Reviewed-on: https://gerrit.libreoffice.org/61473
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>