tdf#119580 sc: Use the regular object popupmenu for charts
Context menu of a chart is supposed to be the same as for
other ole objects, except the additional .uno:ExportAsGraphic.
Given that .uno:ExportAsGraphic is hidden when non-chart
object is selected, we can just place it in the regular ole
object popupmenu (done already), and use that menu also for
charts.

Discussion is in https://gerrit.libreoffice.org/60128 .

Change-Id: I8a07c550998e1db0d2af7f87c625dbd258454bdd
Reviewed-on: https://gerrit.libreoffice.org/61678
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>